week/4 - [Refactor] HobbyView 분리
