+++ SOPT-35-Seminar/SOPT-35-Seminar/Network/View/ChangeHobbyView.swift
@@ -1,0 +1,20 @@
+//
+//  ChangeHobbyView.swift
+//  SOPT-35-Seminar
+//
+//  Created by 최유빈 on 11/16/24.
+//
+
+import UIKit
+
+class ChangeHobbyView: UIView {
+
+    /*
+    // Only override draw() if you perform custom drawing.
+    // An empty implementation adversely affects performance during animation.
+    override func draw(_ rect: CGRect) {
+        // Drawing code
+    }
+    */
+
+}
