+++ SOPT-35-Seminar/SOPT-35-Seminar/Network/ViewController/LoginViewController.swift
@@ -15,6 +15,7 @@ class LoginViewController: UIViewController {
     override func loadView() {
         view = loginView
     }
+    
     override func viewDidLoad() {
         super.viewDidLoad()
